requirements.txt
faster-whisper>=1.0.0
pyannote.audio>=3.1
torch>=2.0
torchaudio>=2.0
pydub>=0.25
noisereduce>=3.0
numpy>=1.24
huggingface_hub>=0.20
scipy>=1.10
soundfile>=0.12
llama-cpp-python>=0.3.0
